requirements.txt
Flask==2.2.5
Pillow==10.0.0
torch>=1.13.0
transformers>=4.30.0
faiss-cpu>=1.7.4
tqdm>=4.65.0
numpy>=1.24.0